filter key can contain floats
Former-commit-id: c488b755f9d5ba4fe63e636c46aad485826d52b7

+++ pybinsim/pose.py
@@ -37,7 +37,8 @@ class Pose:
     @staticmethod
     def from_filterValueList(filter_value_list):
 
-        filter_value_list = np.squeeze(np.asarray(filter_value_list, dtype=np.int16))
+        #filter_value_list = np.squeeze(np.asarray(filter_value_list, dtype=np.int16))
+        filter_value_list = np.squeeze(np.asarray(filter_value_list, dtype=np.float32))
         # format: listener_orientation - listener_position - custom
         if len(filter_value_list) == 9:
             listener_orientation = Orientation(
@@ -85,7 +86,8 @@ class SourcePose:
     @staticmethod
     def from_filterValueList(filter_value_list):
 
-        filter_value_list = np.squeeze(np.asarray(filter_value_list, dtype=np.int16))
+        #filter_value_list = np.squeeze(np.asarray(filter_value_list, dtype=np.int16))
+        filter_value_list = np.squeeze(np.asarray(filter_value_list, dtype=np.float32))
 
         # 'new' format: source_orientation - source_position - custom
         if len(filter_value_list) == 9:
